Move ManualCompletionClient to activity package
